+++ dsigm/mixture/__init__.py
@@ -1,0 +1,4 @@
+from ._sgmm import SGMM
+from ._gmm import GMM
+
+__all__ = ['SGMM', 'GMM']
